+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Fixed unable to update `AmountInput` value after max clicked](https://github.com/multiversx/mx-sdk-dapp-form/pull/147)
+-
 ## [[0.4.14](https://github.com/multiversx/mx-sdk-dapp-form/pull/146)] - 2023-03-08
 - [Upgraded `sdk-dapp` and `sdk-core`](https://github.com/multiversx/mx-sdk-dapp-form/pull/145)